+++ src/main/java/chess/ChessApplication.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.domain.ChessGame;
 import chess.view.InputView;
 import chess.view.OutputView;
